+++ e2e-test/e2e/e2e.go
@@ -21,15 +21,17 @@ import (
 	"os/exec"
 	"time"
 
-	"github.com/onsi/ginkgo"
+	"github.com/onsi/ginkgo/v2"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 	"k8s.io/klog/v2"
+	_ "k8s.io/kubelet"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2edebug "k8s.io/kubernetes/test/e2e/framework/debug"
 	e2ekubectl "k8s.io/kubernetes/test/e2e/framework/kubectl"
-	e2elog "k8s.io/kubernetes/test/e2e/framework/log"
+	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
 	e2epod "k8s.io/kubernetes/test/e2e/framework/pod"
 	utilnet "k8s.io/utils/net"
 
@@ -67,18 +69,18 @@ func setupSuite() {
 				"local-path-storage",
 			})
 		if err != nil {
-			e2elog.Failf("Error deleting orphaned namespaces: %v", err)
+			framework.Failf("Error deleting orphaned namespaces: %v", err)
 		}
 		klog.Infof("Waiting for deletion of the following namespaces: %v", deleted)
 		if err := framework.WaitForNamespacesDeleted(c, deleted, namespaceCleanupTimeout); err != nil {
-			e2elog.Failf("Failed to delete orphaned namespaces %v: %v", deleted, err)
+			framework.Failf("Failed to delete orphaned namespaces %v: %v", deleted, err)
 		}
 	}
 
 	// In large clusters we may get to this point but still have a bunch
 	// of nodes without Routes created. Since this would make a node
 	// unschedulable, we need to wait until all of them are schedulable.
-	framework.ExpectNoError(framework.WaitForAllNodesSchedulable(c, framework.TestContext.NodeSchedulableTimeout))
+	framework.ExpectNoError(e2enode.WaitForAllNodesSchedulable(c, framework.TestContext.NodeSchedulableTimeout))
 
 	//// If NumNodes is not specified then auto-detect how many are scheduleable and not tainted
 	//if framework.TestContext.CloudConfig.NumNodes == framework.DefaultNumNodes {
@@ -95,23 +97,23 @@ func setupSuite() {
 	// wasting the whole run), we allow for some not-ready pods (with the
 	// number equal to the number of allowed not-ready nodes).
 	if err := e2epod.WaitForPodsRunningReady(c, metav1.NamespaceSystem, int32(framework.TestContext.MinStartupPods), int32(framework.TestContext.AllowedNotReadyNodes), podStartupTimeout, map[string]string{}); err != nil {
-		framework.DumpAllNamespaceInfo(c, metav1.NamespaceSystem)
-		e2ekubectl.LogFailedContainers(c, metav1.NamespaceSystem, e2elog.Logf)
-		e2elog.Failf("Error waiting for all pods to be running and ready: %v", err)
+		e2edebug.DumpAllNamespaceInfo(c, metav1.NamespaceSystem)
+		e2ekubectl.LogFailedContainers(c, metav1.NamespaceSystem, framework.Logf)
+		framework.Failf("Error waiting for all pods to be running and ready: %v", err)
 	}
 
 	//if err := framework.WaitForDaemonSets(c, metav1.NamespaceSystem, int32(framework.TestContext.AllowedNotReadyNodes), framework.TestContext.SystemDaemonsetStartupTimeout); err != nil {
-	//	e2elog.Logf("WARNING: Waiting for all daemonsets to be ready failed: %v", err)
+	//	framework.Logf("WARNING: Waiting for all daemonsets to be ready failed: %v", err)
 	//}
 
 	dc := c.DiscoveryClient
 
 	serverVersion, serverErr := dc.ServerVersion()
 	if serverErr != nil {
-		e2elog.Logf("Unexpected server error retrieving version: %v", serverErr)
+		framework.Logf("Unexpected server error retrieving version: %v", serverErr)
 	}
 	if serverVersion != nil {
-		e2elog.Logf("kube-apiserver version: %s", serverVersion.GitVersion)
+		framework.Logf("kube-apiserver version: %s", serverVersion.GitVersion)
 	}
 }
 
@@ -151,7 +153,7 @@ func setupSuitePerGinkgoNode() {
 		klog.Fatal("Error loading client: ", err)
 	}
 	framework.TestContext.IPFamily = getDefaultClusterIPFamily(c)
-	e2elog.Logf("Cluster IP family: %s", framework.TestContext.IPFamily)
+	framework.Logf("Cluster IP family: %s", framework.TestContext.IPFamily)
 }
 
 // getDefaultClusterIPFamily obtains the default IP family of the cluster
@@ -163,7 +165,7 @@ func getDefaultClusterIPFamily(c kubernetes.Interface) string {
 	// Get the ClusterIP of the kubernetes service created in the default namespace
 	svc, err := c.CoreV1().Services(metav1.NamespaceDefault).Get(context.TODO(), "kubernetes", metav1.GetOptions{})
 	if err != nil {
-		e2elog.Failf("Failed to get kubernetes service ClusterIP: %v", err)
+		framework.Failf("Failed to get kubernetes service ClusterIP: %v", err)
 	}
 
 	if utilnet.IsIPv6String(svc.Spec.ClusterIP) {