+++ pkg/dashboard/store/event/event_test.go
@@ -23,7 +23,7 @@ import (
 
 	sqlmock "github.com/DATA-DOG/go-sqlmock"
 	"github.com/jinzhu/gorm"
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"github.com/chaos-mesh/chaos-mesh/pkg/dashboard/core"